[refactor] es dummy api 모듈 분리
